refactor(v5)!: settle transaction, tracing and paging options
Nine in ten WithTx calls passed nil in the middle for a TxOptions struct, the
only struct-shaped option left beside RuntimeOption and BatchOption. Tracers
got an operation name but no table, so a span could only say "update".
HTTP paging took three steps and made handlers pass runtime.MaxPageSize()
themselves, although Page already caps the size. Query.String rendered
SQLite whatever the target, and SearchTerm named the full-text side after
the keyword-search side.

- WithTx(ctx, fn, ...TxOption) / WithTxResult(ctx, fn, ...TxOption) with
  WithIsolation, WithReadOnly, WithRetry and WithRetryPolicy; TxOptions is
  gone and DefaultRetryPolicy returns a value.
- Tracer takes TraceInfo{Op, Table}: the written table or the FROM table.
  UpdateTable/DeleteFrom trace as update/delete; TraceOpScalar and
  TraceOpExec are gone.
- PageRequest.Paging and Keyset validate the request; Validate, Normalize
  and Runtime.MaxPageSize are gone, Page caps the size and reports it.
- Query.String is gone; SearchTerm is MatchTerm; OptimisticLockError and
  RowStateError report Expected as int64.
- WithTablePolicy/WithIndexPolicy stay: the academy example manages indexes
  on tables mock.sql creates, which one merged option cannot say.

Verified with make harness and go test ./...; trace_test.go checks the table
each entry point reports, paging_test.go the new validation and the cap.

BREAKING CHANGE: WithTx takes options after the callback, Tracer takes a
TraceInfo, and PageRequest has no Validate/Normalize.

+++ errors.go
@@ -13,7 +13,7 @@ type RowStateError struct {
 	Table    string
 	Op       string
 	Need     string
-	Expected int
+	Expected int64
 	Actual   int64
 }
 
@@ -32,10 +32,10 @@ func IsRowStateError(err error) bool {
 // OptimisticLockError reports that a version-guarded write matched fewer rows than
 // it was given: another writer changed or deleted one of them first. It is a
 // business outcome to handle, typically by reloading and retrying; see
-// TxOptions.RetryIf and IsOptimisticLockError.
+// WithRetry and IsOptimisticLockError.
 type OptimisticLockError struct {
 	Table    string
-	Expected int
+	Expected int64
 	Actual   int64
 }
